:hammer: 생성한 뷰홀더로 코드 적용
